+++ amp-toolbox-linter/src/index.ts
@@ -8,27 +8,27 @@ import { isStatusOk, isJson, isAccessControlHeaders } from "./filter";
 import { absoluteUrl, fetchToCurl } from "./url";
 import { ampType } from "./helper";
 import { cli } from "./cli";
-import { SchemaMetadataIsNews } from "./rule/SchemaMetadataIsNews";
-import { LinkRelCanonicalIsOk } from "./rule/LinkRelCanonicalIsOk";
-import { AmpVideoIsSmall } from "./rule/AmpVideoIsSmall";
-import { BookendExists } from "./rule/BookendExists";
-import { BookendAppearsOnOrigin } from "./rule/BookendAppearsOnOrigin";
-import { BookendAppearsOnCache } from "./rule/BookendAppearsOnCache";
-import { AmpVideoIsSpecifiedByAttribute } from "./rule/AmpVideoIsSpecifiedByAttribute";
-import { StoryRuntimeIsV1 } from "./rule/StoryRuntimeIsV1";
-import { StoryMetadataIsV1 } from "./rule/StoryMetadataIsV1";
-import { MetaCharsetIsFirst } from "./rule/MetaCharsetIsFirst";
-import { RuntimeIsPreloaded } from "./rule/RuntimeIsPreloaded";
-import { StoryIsMostlyText } from "./rule/StoryIsMostlyText";
-import { StoryMetadataThumbnailsAreOk } from "./rule/StoryMetadataThumbnailsAreOk";
-import { AmpImgHeightWidthIsOk } from "./rule/AmpImgHeightWidthIsOk";
-import { AmpImgAmpPixelPreferred } from "./rule/AmpImgAmpPixelPreferred";
-import { EndpointsAreAccessibleFromOrigin } from "./rule/EndpointsAreAccessibleFromOrigin";
-import { EndpointsAreAccessibleFromCache } from "./rule/EndpointsAreAccessibleFromCache";
-import { SxgVaryOnAcceptAct } from "./rule/SxgVaryOnAcceptAct";
-import { SxgContentNegotiationIsOk } from "./rule/SxgContentNegotiationIsOk";
-import { SxgDumpSignedExchangeVerify } from "./rule/SxgDumpSignedExchangeVerify";
-import { SxgAmppkgIsForwarded } from "./rule/SxgAmppkgIsForwarded";
+import { SchemaMetadataIsNews } from "./rules/SchemaMetadataIsNews";
+import { LinkRelCanonicalIsOk } from "./rules/LinkRelCanonicalIsOk";
+import { AmpVideoIsSmall } from "./rules/AmpVideoIsSmall";
+import { BookendExists } from "./rules/BookendExists";
+import { BookendAppearsOnOrigin } from "./rules/BookendAppearsOnOrigin";
+import { BookendAppearsOnCache } from "./rules/BookendAppearsOnCache";
+import { AmpVideoIsSpecifiedByAttribute } from "./rules/AmpVideoIsSpecifiedByAttribute";
+import { StoryRuntimeIsV1 } from "./rules/StoryRuntimeIsV1";
+import { StoryMetadataIsV1 } from "./rules/StoryMetadataIsV1";
+import { MetaCharsetIsFirst } from "./rules/MetaCharsetIsFirst";
+import { RuntimeIsPreloaded } from "./rules/RuntimeIsPreloaded";
+import { StoryIsMostlyText } from "./rules/StoryIsMostlyText";
+import { StoryMetadataThumbnailsAreOk } from "./rules/StoryMetadataThumbnailsAreOk";
+import { AmpImgHeightWidthIsOk } from "./rules/AmpImgHeightWidthIsOk";
+import { AmpImgAmpPixelPreferred } from "./rules/AmpImgAmpPixelPreferred";
+import { EndpointsAreAccessibleFromOrigin } from "./rules/EndpointsAreAccessibleFromOrigin";
+import { EndpointsAreAccessibleFromCache } from "./rules/EndpointsAreAccessibleFromCache";
+import { SxgVaryOnAcceptAct } from "./rules/SxgVaryOnAcceptAct";
+import { SxgContentNegotiationIsOk } from "./rules/SxgContentNegotiationIsOk";
+import { SxgDumpSignedExchangeVerify } from "./rules/SxgDumpSignedExchangeVerify";
+import { SxgAmppkgIsForwarded } from "./rules/SxgAmppkgIsForwarded";
 import { Rule, RuleConstructor } from "./rule";
 
 export enum LintType {
