+++ amp-toolbox-linter/src/rules/AmpImgAmpPixelPreferred.ts
