+++ amp-toolbox-linter/src/rules/AmpImgHeightWidthIsOk.ts
