+++ amp-toolbox-linter/src/rules/AmpVideoIsSmall.ts
