+++ amp-toolbox-linter/src/rules/AmpVideoIsSpecifiedByAttribute.ts
