+++ amp-toolbox-linter/src/rules/BookendAppearsOnCache.ts
