+++ amp-toolbox-linter/src/rules/BookendAppearsOnOrigin.ts
