+++ amp-toolbox-linter/src/rules/EndpointsAreAccessibleFromCache.ts
