+++ amp-toolbox-linter/src/rules/EndpointsAreAccessibleFromOrigin.ts
