+++ amp-toolbox-linter/src/rules/IsValid.ts
