+++ amp-toolbox-linter/src/rules/LinkRelCanonicalIsOk.ts
