+++ amp-toolbox-linter/src/rules/SchemaMetadataIsNews.ts
