+++ amp-toolbox-linter/src/rules/StoryIsMostlyText.ts
