+++ amp-toolbox-linter/src/rules/StoryMetadataIsV1.ts
