+++ amp-toolbox-linter/src/rules/StoryMetadataThumbnailsAreOk.ts
