+++ amp-toolbox-linter/src/rules/SxgAmppkgIsForwarded.ts
