+++ amp-toolbox-linter/src/rules/SxgContentNegotiationIsOk.ts
