+++ amp-toolbox-linter/src/rules/SxgVaryOnAcceptAct.ts
